Rename `TrackRequest` and `TrackRequestOptions` to `DownloadRequest` and `DownloadRequestOptions`
